+++ Kernel/Devices/Storage/AHCI/InterruptHandler.h
@@ -7,8 +7,8 @@
 #pragma once
 
 #include <Kernel/Devices/Device.h>
-#include <Kernel/Devices/Storage/ATA/AHCI/Controller.h>
-#include <Kernel/Devices/Storage/ATA/AHCI/Port.h>
+#include <Kernel/Devices/Storage/AHCI/Controller.h>
+#include <Kernel/Devices/Storage/AHCI/Port.h>
 #include <Kernel/Devices/Storage/StorageDevice.h>
 #include <Kernel/Interrupts/PCIIRQHandler.h>
 #include <Kernel/Library/LockRefPtr.h>
