--- Kernel/Devices/Storage/ATA/ATAController.h
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2021, Liav A. <liavalb@hotmail.co.il>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#pragma once
-
-#include <AK/OwnPtr.h>
-#include <AK/Types.h>
-#include <Kernel/Devices/BlockDevice.h>
-#include <Kernel/Devices/Storage/StorageController.h>
-#include <Kernel/Library/LockRefPtr.h>
-
-namespace Kernel {
-
-class AsyncBlockDeviceRequest;
-
-class ATADevice;
-class ATAController
-    : public StorageController
-    , public LockWeakable<ATAController> {
-public:
-    virtual void start_request(ATADevice const&, AsyncBlockDeviceRequest&) = 0;
-
-protected:
-    ATAController();
-};
-}
