+++ Meta/gn/secondary/Kernel/BUILD.gn
@@ -319,7 +319,6 @@ source_set("arch_sources") {
       "Arch/x86_64/I8042Reboot.cpp",
       "Arch/x86_64/ISABus/HID/VMWareMouseDevice.cpp",
       "Arch/x86_64/ISABus/I8042Controller.cpp",
-      "Arch/x86_64/ISABus/IDEController.cpp",
       "Arch/x86_64/ISABus/SerialDevice.cpp",
       "Arch/x86_64/InterruptEntry.cpp",
       "Arch/x86_64/InterruptManagement.cpp",
@@ -328,7 +327,6 @@ source_set("arch_sources") {
       "Arch/x86_64/Interrupts/IOAPIC.cpp",
       "Arch/x86_64/Interrupts/PIC.cpp",
       "Arch/x86_64/PCI/Controller/HostBridge.cpp",
-      "Arch/x86_64/PCI/IDELegacyModeController.cpp",
       "Arch/x86_64/PCI/Initializer.cpp",
       "Arch/x86_64/PCI/MSI.cpp",
       "Arch/x86_64/PCSpeaker.cpp",
@@ -556,15 +554,12 @@ executable("Kernel_bin") {
     "Devices/KCOVInstance.cpp",
     "Devices/PCISerialDevice.cpp",
     "Devices/SerialDevice.cpp",
-    "Devices/Storage/ATA/AHCI/Controller.cpp",
-    "Devices/Storage/ATA/AHCI/InterruptHandler.cpp",
-    "Devices/Storage/ATA/AHCI/Port.cpp",
-    "Devices/Storage/ATA/ATAController.cpp",
-    "Devices/Storage/ATA/ATADevice.cpp",
-    "Devices/Storage/ATA/ATADiskDevice.cpp",
-    "Devices/Storage/ATA/ATAPort.cpp",
-    "Devices/Storage/ATA/GenericIDE/Channel.cpp",
-    "Devices/Storage/ATA/GenericIDE/Controller.cpp",
+    "Devices/Storage/AHCI/ATAController.cpp",
+    "Devices/Storage/AHCI/ATADevice.cpp",
+    "Devices/Storage/AHCI/ATADiskDevice.cpp",
+    "Devices/Storage/AHCI/Controller.cpp",
+    "Devices/Storage/AHCI/InterruptHandler.cpp",
+    "Devices/Storage/AHCI/Port.cpp",
     "Devices/Storage/DiskPartition.cpp",
     "Devices/Storage/NVMe/NVMeController.cpp",
     "Devices/Storage/NVMe/NVMeInterruptQueue.cpp",
